+++ packages/failurechecker/src/failurechecker.ts
@@ -17,16 +17,15 @@
 // check whether type bindings are already published.
 
 // eslint-disable-next-line node/no-extraneous-import
-import {Application} from 'probot';
-import {Octokit} from '@octokit/rest';
-import {IssuesListForRepoResponseData} from '@octokit/types';
-type OctokitType = InstanceType<typeof Octokit>;
-
+import {Application, ProbotOctokit} from 'probot';
 import {logger} from 'gcf-utils';
+
+type OctokitType = InstanceType<typeof ProbotOctokit>;
 
 // labels indicative of the fact that a release has not completed yet.
 const RELEASE_LABELS = ['autorelease: pending', 'autorelease: failed'];
 const RELEASE_TYPE_NO_PUBLISH = ['go-yoshi'];
+const SUCCESSFUL_PUBLISH_LABEL = 'autorelease: published';
 
 // We open an issue that a release has failed if it's been longer than 3
 // hours and we're within normal working hours.
@@ -42,9 +41,16 @@ interface ConfigurationOptions {
   releaseType?: string;
 }
 
+// exported for testing purposes
+export const TimeMethods = {
+  Date: () => {
+    return new Date();
+  },
+};
+
 export function failureChecker(app: Application) {
   app.on('schedule.repository' as '*', async context => {
-    const utcHour = new Date().getUTCHours();
+    const utcHour = TimeMethods.Date().getUTCHours();
     const owner = context.payload.organization.login;
     const repo = context.payload.repository.name;
 
@@ -66,7 +72,8 @@ export function failureChecker(app: Application) {
       labels.push('autorelease: tagged');
     }
 
-    const now = new Date().getTime();
+    const now = TimeMethods.Date().getTime();
+    const failed: number[] = [];
     for (const label of labels) {
       const results = (
         await context.github.issues.listForRepo({
@@ -94,31 +101,39 @@ export function failureChecker(app: Application) {
               pull_number: issue.number,
             })
           ).data;
-          if (pr.merged_at) {
-            await openWarningIssue(owner, repo, pr.number, context.github);
+          if (
+            pr.merged_at &&
+            !pr.labels.some(l => l.name === SUCCESSFUL_PUBLISH_LABEL)
+          ) {
+            failed.push(pr.number);
           }
         }
       }
     }
+
+    await manageWarningIssue(owner, repo, failed, context.github);
     logger.info(`it's alive! event for ${repo}`);
   });
 
+  function buildIssueBody(prNumbers: number[]): string {
+    const list = prNumbers.map(prNumber => `* #${prNumber}`).join('\n');
+    return `The following release PRs may have failed:\n\n${list}`;
+  }
+
   const ISSUE_TITLE = 'Warning: a recent release failed';
   const LABELS = 'type: process';
-  async function openWarningIssue(
+  async function manageWarningIssue(
     owner: string,
     repo: string,
-    prNumber: number,
+    prNumbers: number[],
     github: OctokitType
   ) {
-    const issues = (
-      await github.issues.listForRepo({
-        owner,
-        repo,
-        labels: LABELS,
-        per_page: 32,
-      })
-    ).data as IssuesListForRepoResponseData;
+    const {data: issues} = await github.issues.listForRepo({
+      owner,
+      repo,
+      labels: LABELS,
+      per_page: 32,
+    });
     const warningIssue = issues.find(issue => {
       return issue.title.includes(ISSUE_TITLE);
     });
@@ -127,16 +142,43 @@ export function failureChecker(app: Application) {
     // a cron effects our usage limits:
     logger.info((await github.rateLimit.get()).data);
 
-    if (warningIssue) {
-      logger.info(`a warning issue was already opened for pr ${prNumber}`);
+    // existing issue and no failures - close the existing issue
+    if (prNumbers.length === 0) {
+      if (warningIssue) {
+        await github.issues.update({
+          owner,
+          repo,
+          issue_number: warningIssue.number,
+          state: 'closed',
+        });
+      }
       return;
     }
-    app.log(`opening warning issue on ${repo} for PR #${prNumber}`);
-    return github.issues.create({
+
+    const body = buildIssueBody(prNumbers);
+
+    if (warningIssue) {
+      if (warningIssue.body === body) {
+        // issue already up-to-date
+        logger.info(`a warning issue was already opened for prs ${prNumbers}`);
+        return;
+      }
+
+      // Update the existing issue
+      await github.issues.update({
+        owner,
+        repo,
+        issue_number: warningIssue.number,
+        body,
+      });
+      return;
+    }
+    app.log(`opening warning issue on ${repo} for PR #${prNumbers}`);
+    await github.issues.create({
       owner,
       repo,
       title: ISSUE_TITLE,
-      body: `The release PR #${prNumber} is still in a pending state after several hours`,
+      body,
       labels: LABELS.split(','),
     });
   }