+++ packages/failurechecker/test/failurechecker.ts
@@ -13,16 +13,12 @@
 // limitations under the License.
 
 // eslint-disable-next-line node/no-extraneous-import
-import {Probot, createProbot} from 'probot';
+import {Probot, createProbot, ProbotOctokit} from 'probot';
 import snapshot from 'snap-shot-it';
 import nock from 'nock';
 import {describe, it, beforeEach, afterEach} from 'mocha';
 import * as sinon from 'sinon';
-import {failureChecker} from '../src/failurechecker';
-// eslint-disable-next-line node/no-extraneous-import
-import {Octokit} from '@octokit/rest';
-import {config} from '@probot/octokit-plugin-config';
-const TestingOctokit = Octokit.plugin(config);
+import {failureChecker, TimeMethods} from '../src/failurechecker';
 
 nock.disableNetConnect();
 
@@ -30,20 +26,22 @@ describe('failurechecker', () => {
   let probot: Probot;
 
   beforeEach(() => {
-    sinon.useFakeTimers(new Date(Date.UTC(2020, 1, 1, 20)));
+    // Later versions of probot seem to have problems dealing with fake sinon timers.
+    sinon.stub(TimeMethods, 'Date').returns(new Date(Date.UTC(2020, 1, 1, 20)));
     probot = createProbot({
       githubToken: 'abc123',
-      Octokit: TestingOctokit as any,
+      Octokit: ProbotOctokit.defaults({
+        retry: {enabled: false},
+        throttle: {enabled: false},
+      }),
     });
-
     probot.load(failureChecker);
   });
 
   afterEach(() => {
     nock.cleanAll();
+    sinon.restore();
   });
-
-  afterEach(() => sinon.restore());
 
   it('opens an issue on GitHub if there exists a pending label > threshold', async () => {
     const requests = nock('https://api.github.com')
@@ -69,6 +67,7 @@ describe('failurechecker', () => {
       .reply(200, {
         number: 33,
         merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: pending'}],
       })
       .get(
         '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20tagged&state=closed&sort=updated&direction=desc&per_page=16'
@@ -91,7 +90,7 @@ describe('failurechecker', () => {
       .reply(200);
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -140,6 +139,7 @@ describe('failurechecker', () => {
       .reply(200, {
         number: 33,
         merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: pending'}],
       })
       .get(
         '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
@@ -154,7 +154,7 @@ describe('failurechecker', () => {
       .reply(200);
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -189,10 +189,16 @@ describe('failurechecker', () => {
       .get(
         '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20failed&state=closed&sort=updated&direction=desc&per_page=16'
       )
+      .reply(200, [])
+      .get('/rate_limit')
+      .reply(200, {})
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
       .reply(200, []);
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -237,10 +243,16 @@ describe('failurechecker', () => {
           number: 33,
           updated_at: date,
         },
-      ]);
+      ])
+      .get('/rate_limit')
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, []);
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -284,10 +296,16 @@ describe('failurechecker', () => {
           number: 33,
           updated_at: '2020-01-25T13:33:48Z',
         },
-      ]);
+      ])
+      .get('/rate_limit')
+      .reply(200, {})
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, []);
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -336,6 +354,7 @@ describe('failurechecker', () => {
       .reply(200, {
         number: 33,
         merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: pending'}],
       })
       .get(
         '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
@@ -343,13 +362,15 @@ describe('failurechecker', () => {
       .reply(200, [
         {
           title: 'Warning: a recent release failed',
+          number: 44,
+          body: 'The following release PRs may have failed:\n\n* #33',
         },
       ])
       .get('/rate_limit')
       .reply(200, {});
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -368,10 +389,11 @@ describe('failurechecker', () => {
 
   it('does not run outside of working hours', async () => {
     // UTC 05:00 is outside of working hours:
-    sinon.useFakeTimers(new Date(Date.UTC(2020, 1, 1, 5)));
+    sinon.restore();
+    sinon.stub(TimeMethods, 'Date').returns(new Date(Date.UTC(2020, 1, 1, 5)));
 
     await probot.receive({
-      name: 'schedule.repository' as any,
+      name: 'schedule.repository' as '*',
       payload: {
         repository: {
           name: 'nodejs-foo',
@@ -382,5 +404,293 @@ describe('failurechecker', () => {
       },
       id: 'abc123',
     });
+  });
+
+  it('closes a PR once failing releases are handled', async () => {
+    const requests = nock('https://api.github.com')
+      .get('/repos/bcoe/nodejs-foo/contents/.github%2Frelease-please.yml')
+      .reply(
+        200,
+        Buffer.from(
+          JSON.stringify({
+            releaseType: 'node',
+          })
+        )
+      )
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20pending&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [
+        {
+          number: 33,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+      ])
+      .get('/repos/googleapis/nodejs-foo/pulls/33')
+      .reply(200, {
+        number: 33,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [
+          {name: 'autorelease: failed'},
+          {name: 'autorelease: published'},
+        ],
+      })
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20tagged&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20failed&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, [
+        {
+          title: 'Warning: a recent release failed',
+          number: 44,
+        },
+      ])
+      .get('/rate_limit')
+      .reply(200, {})
+      .patch('/repos/googleapis/nodejs-foo/issues/44', body => {
+        snapshot(body);
+        return true;
+      })
+      .reply(200, []);
+
+    await probot.receive({
+      name: 'schedule.repository' as '*',
+      payload: {
+        repository: {
+          name: 'nodejs-foo',
+          owner: {
+            login: 'bcoe',
+          },
+        },
+        organization: {
+          login: 'googleapis',
+        },
+      },
+      id: 'abc123',
+    });
+    requests.done();
+  });
+
+  it('ignores a PR both failed and published', async () => {
+    const requests = nock('https://api.github.com')
+      .get('/repos/bcoe/nodejs-foo/contents/.github%2Frelease-please.yml')
+      .reply(
+        200,
+        Buffer.from(
+          JSON.stringify({
+            releaseType: 'node',
+          })
+        )
+      )
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20pending&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [
+        {
+          number: 33,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+      ])
+      .get('/repos/googleapis/nodejs-foo/pulls/33')
+      .reply(200, {
+        number: 33,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [
+          {name: 'autorelease: failed'},
+          {name: 'autorelease: published'},
+        ],
+      })
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20tagged&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20failed&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, [])
+      .get('/rate_limit')
+      .reply(200, {});
+
+    await probot.receive({
+      name: 'schedule.repository' as '*',
+      payload: {
+        repository: {
+          name: 'nodejs-foo',
+          owner: {
+            login: 'bcoe',
+          },
+        },
+        organization: {
+          login: 'googleapis',
+        },
+      },
+      id: 'abc123',
+    });
+    requests.done();
+  });
+
+  it('updates an issue with new failures', async () => {
+    const requests = nock('https://api.github.com')
+      .get('/repos/bcoe/nodejs-foo/contents/.github%2Frelease-please.yml')
+      .reply(
+        200,
+        Buffer.from(
+          JSON.stringify({
+            releaseType: 'node',
+          })
+        )
+      )
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20pending&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [
+        {
+          number: 33,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+        {
+          number: 34,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+      ])
+      .get('/repos/googleapis/nodejs-foo/pulls/33')
+      .reply(200, {
+        number: 33,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: failed'}],
+      })
+      .get('/repos/googleapis/nodejs-foo/pulls/34')
+      .reply(200, {
+        number: 34,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: failed'}],
+      })
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20tagged&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20failed&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, [
+        {
+          title: 'Warning: a recent release failed',
+          number: 44,
+          body: 'The following release PRs may have failed:\n\n* #33',
+        },
+      ])
+      .get('/rate_limit')
+      .reply(200, {})
+      .patch('/repos/googleapis/nodejs-foo/issues/44', body => {
+        snapshot(body);
+        return true;
+      })
+      .reply(200, []);
+
+    await probot.receive({
+      name: 'schedule.repository' as '*',
+      payload: {
+        repository: {
+          name: 'nodejs-foo',
+          owner: {
+            login: 'bcoe',
+          },
+        },
+        organization: {
+          login: 'googleapis',
+        },
+      },
+      id: 'abc123',
+    });
+    requests.done();
+  });
+
+  it('opens an issue with multiple failures', async () => {
+    const requests = nock('https://api.github.com')
+      .get('/repos/bcoe/nodejs-foo/contents/.github%2Frelease-please.yml')
+      .reply(
+        200,
+        Buffer.from(
+          JSON.stringify({
+            releaseType: 'node',
+          })
+        )
+      )
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20pending&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [
+        {
+          number: 33,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+        {
+          number: 34,
+          updated_at: '2020-01-30T13:33:48Z',
+        },
+      ])
+      .get('/repos/googleapis/nodejs-foo/pulls/33')
+      .reply(200, {
+        number: 33,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: failed'}],
+      })
+      .get('/repos/googleapis/nodejs-foo/pulls/34')
+      .reply(200, {
+        number: 34,
+        merged_at: '2020-01-30T13:33:48Z',
+        labels: [{name: 'autorelease: failed'}],
+      })
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20tagged&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=autorelease%3A%20failed&state=closed&sort=updated&direction=desc&per_page=16'
+      )
+      .reply(200, [])
+      .get(
+        '/repos/googleapis/nodejs-foo/issues?labels=type%3A%20process&per_page=32'
+      )
+      .reply(200, [])
+      .get('/rate_limit')
+      .reply(200, {})
+      .post('/repos/googleapis/nodejs-foo/issues', body => {
+        snapshot(body);
+        return true;
+      })
+      .reply(200, []);
+
+    await probot.receive({
+      name: 'schedule.repository' as '*',
+      payload: {
+        repository: {
+          name: 'nodejs-foo',
+          owner: {
+            login: 'bcoe',
+          },
+        },
+        organization: {
+          login: 'googleapis',
+        },
+      },
+      id: 'abc123',
+    });
+    requests.done();
   });
 });